requirements.txt
Flask==2.0.1
Flask-Login==0.6.3
Werkzeug==2.0.3
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
oauthlib==3.2.2
google-auth-oauthlib==1.2.0
gunicorn==23.0.0
flask-cors==4.0.0
